+++ index.html
@@ -4,8 +4,11 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
-    <script src="https://dapi.kakao.com/v2/maps/sdk.js?appkey=bcc85e56b4e6da53a218658cd336c213&libraries=services,clusterer,drawing"></script>
+    <title>Walkie-Paw</title>
+    <script
+        type="text/javascript"
+        src="https://dapi.kakao.com/v2/maps/sdk.js?appkey=39540ea4b1e5f4b36c10411338099411&libraries=services,clusterer,drawing"
+    ></script>
     <!-- 지도 API는 반드시 실행 코드보다 먼저 실행되어야 한다. head, body 어느 곳에 넣든 상관 없음. -->
   </head>
   <body>
