+++ package.json
@@ -43,7 +43,7 @@
 		"eslint-plugin-react-refresh": "^0.4.6",
 		"postcss": "^8.4.38",
 		"tailwindcss": "^3.4.4",
-		"vite": "^5.2.0",
+		"vite": "^5.3.1",
 		"vite-plugin-eslint": "^1.8.1"
 	}
 }
