+++ src/pages/Login.jsx
@@ -1,5 +1,3 @@
-// Login.js
-
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import UserInput from '../components/UserInput';
@@ -54,10 +52,10 @@ const Login = () => {
     setErrors(newErrors);
   };
 
-  const isInvaild =
-    userInfo.email.includes('@') &&
-    userInfo.email.includes('.') &&
-    userInfo.password.length >= 8;
+  const isInvalid =
+      userInfo.email.includes('@') &&
+      userInfo.email.includes('.') &&
+      userInfo.password.length >= 8;
 
   const handleKakaoLogin = () => {
     window.Kakao.Auth.login({
@@ -79,64 +77,84 @@ const Login = () => {
     console.log('Google login clicked');
   };
 
-  return (
-    <div style={{ maxWidth: '1200px', margin: '0 auto' }}>
-      <div className="flex flex-col items-center py-8">
-        <img className="h-48" src={pawpaw} alt="발바닥로고" />
-        <UserInput
-          type="text"
-          placeholder="아이디/이메일"
-          value={userInfo.email}
-          name="email"
-          className="w-3/4 px-3 py-2 border-2 border-gray-400 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-400 mb-4"
-          onChange={handleInputChange}
-        />
-        <UserInput
-          type="password"
-          placeholder="비밀번호"
-          value={userInfo.password}
-          name="password"
-          className="w-3/4 px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-400 mb-6"
-          onChange={handleInputChange}
-        />
-        <UserButton
-          text="로그인"
-          disabled={!isInvaild}
-          className={`w-3/4 py-3 rounded-lg transition-colors duration-300 ${
-            isInvaild
-              ? 'bg-gradient-to-r from-blue-500 to-blue-700 hover:from-blue-600 hover:to-blue-800 text-white'
-              : 'bg-gray-300 cursor-not-allowed'
-          } mb-4 font-semibold shadow-md`}
-        />
-        <div className="flex justify-between w-3/4 mb-6">
-          <button className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300">
-            아이디 찾기
-          </button>
-          <button className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300">
-            비밀번호 찾기
-          </button>
-          <button
-            type="button"
-            className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300"
-            onClick={() => navigate('/signupform')}
-          >
-            회원가입
-          </button>
-        </div>
+  const onSubmit = () => {
+    // 여기에 실제 로그인 로직 구현
+    console.log('Login submit');
+  };
 
-        <div className="flex justify-between w-3/4 mb-4">
-          <button onClick={handleKakaoLogin}>
-            <img src={kakao_login_medium_narrow} alt="kakao" />
-          </button>
-          <button onClick={handleNaverLogin} style={{ width: '183px', height: '45px' }}>
-            <img src={naver_login} alt="naver" style={{ width: '100%', height: '100%' }} />
-          </button>
-          <button onClick={handleGoogleLogin} style={{ width: '183px', height: '45px' }}>
-            <img src={google_login} alt="google" style={{ width: '100%', height: '100%' }} />
-          </button>
+  return (
+      <div style={{ maxWidth: '1200px', margin: '0 auto' }}>
+        <div className="flex flex-col items-center py-8">
+          <img className="h-72" src={pawpaw} alt="발바닥로고" />
+          <UserInput
+              type="text"
+              placeholder="아이디/이메일"
+              value={userInfo.email}
+              name="email"
+              className={`w-3/4 px-3 py-2 border-2 border-gray-400 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-400 mb-4 ${
+                  errors.email ? 'border-red-500' : ''
+              }`}
+              onChange={handleInputChange}
+          />
+          {errors.email && (
+              <span className="text-sm text-red-500">{errors.email}</span>
+          )}
+          <UserInput
+              type="password"
+              placeholder="비밀번호"
+              value={userInfo.password}
+              name="password"
+              className={`w-3/4 px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-400 mb-6 ${
+                  errors.password ? 'border-red-500' : ''
+              }`}
+              onChange={handleInputChange}
+          />
+          {errors.password && (
+              <span className="text-sm text-red-500">{errors.password}</span>
+          )}
+          <UserButton
+              text="로그인"
+              disabled={!isInvalid}
+              className={`w-3/4 py-3 rounded-lg transition-colors duration-300 ${
+                  isInvalid
+                      ? 'bg-gradient-to-r from-blue-500 to-blue-700 hover:from-blue-600 hover:to-blue-800 text-white'
+                      : 'bg-gray-300 cursor-not-allowed'
+              } mb-4 font-semibold shadow-md`}
+              onClick={onSubmit}
+          />
+          <div className="flex justify-between w-3/4 mb-6">
+            <button className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300">
+              아이디 찾기
+            </button>
+            <button className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300">
+              비밀번호 찾기
+            </button>
+            <button
+                type="button"
+                className="text-sm text-gray-600 hover:text-blue-500 transition-colors duration-300"
+                onClick={() => navigate('/signupform')}
+            >
+              회원가입
+            </button>
+          </div>
+
+          <div className="flex justify-between w-3/4 mb-4">
+            <button onClick={handleKakaoLogin}>
+              <img
+                  src={kakao_login_medium_narrow}
+                  alt="kakao"
+                  style={{ width: 'auto', height: 'auto' }}
+              />
+            </button>
+            <button onClick={handleNaverLogin} style={{ width: '183px', height: '45px' }}>
+              <img src={naver_login} alt="naver" style={{ width: '100%', height: '100%' }} />
+            </button>
+            <button onClick={handleGoogleLogin} style={{ width: '183px', height: '45px' }}>
+              <img src={google_login} alt="google" style={{ width: '100%', height: '100%' }} />
+            </button>
+          </div>
         </div>
       </div>
-    </div>
   );
 };
