+++ src/pages/OAuth/SignUpForm.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import styled from 'styled-components';
+import AuthTemplate from "../../components/auth/AuthTemplate";
+
+
+const SignUpForm = () => {
+
+
+  return (
+      <AuthTemplate>
+
+      </AuthTemplate>
+  );
+};
+
+export default SignUpForm;
